Extracted fetch logic to helper file.
