+++ assets/js/components/Dashboard/Dashboard.tsx
@@ -7,16 +7,8 @@ import { Place } from "../../models/place";
 import { ScreensByAlert } from "../../models/screensByAlert";
 import Sidebar from "./Sidebar";
 import { useInterval } from "../../hooks/useInterval";
+import { fetchAlerts, fetchPlaces } from "../../utils/api";
 
-interface AlertsResponse {
-  alerts: Alert[];
-  screens_by_alert: ScreensByAlert;
-}
-
-interface AlertsResponse {
-  alerts: Alert[];
-  screens_by_alert: ScreensByAlert;
-}
 type ContextType = {
   places: Place[];
   alerts: Alert[];
@@ -30,22 +22,22 @@ const Dashboard: ComponentType = () => {
     {}
   );
 
-  useInterval(() => {
-    fetch("/api/alerts")
-      .then((response) => response.json())
-      .then(({ alerts, screens_by_alert }: AlertsResponse) => {
-        setAlerts(alerts);
-        setScreensByAlertMap(screens_by_alert);
-      });
-  }, 4000);
-
   useEffect(() => {
-    fetch("/api/dashboard")
-      .then((response) => response.json())
-      .then((placeList: []) => {
-        setPlaces(placeList);
-      });
+    fetchPlaces().then((placesList) => setPlaces(placesList));
+
+    fetchAlerts().then(({ alerts, screens_by_alert }) => {
+      setAlerts(alerts);
+      setScreensByAlertMap(screens_by_alert);
+    });
   }, []);
+
+  // Fetch alerts every 4 seconds.
+  useInterval(() => {
+    fetchAlerts().then(({ alerts, screens_by_alert }) => {
+      setAlerts(alerts);
+      setScreensByAlertMap(screens_by_alert);
+    });
+  }, 4000);
 
   return (
     <div className="screenplay-container">
